Making keep(while: ), take(while: ) and skip(while:) array extensions, collection extensions.
